requirements.txt
Django==2.1.7
django-cors-headers==2.5.2
djangorestframework==3.9.2
djangorestframework-gis==0.14
pytz==2018.9
psycopg2-binary==2.7.7
gunicorn==19.9.0
Fiona==1.8.6
flake8==3.7.7
requests==2.22.0
gsconfig-py3==1.0.7
urllib3==1.25.3
django-model-utils==3.1.2
